chore: Move parser into separate file
